requirements.txt
oci[adk]>=2.160.0
fastapi>=0.104.1
uvicorn[standard]>=0.37.0
websockets==12.0
pydantic>=2.11.0
python-multipart>=0.0.20
python-dotenv==1.0.0
requests==2.31.0

# Google ADK Dependencies for Ticket Creation Agent
google-cloud-aiplatform>=1.60.0
google-auth>=2.25.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.1
google-cloud-core>=2.4.1
openai>=1.50.0
